sync

+++ prosty/examples/limero.cpp
@@ -5,6 +5,7 @@
 #include <optional>
 #include <ArduinoJson.h>
 #include <actor.h>
+#include <result.h>
 
 
 typedef std::vector<uint8_t> Bytes;
@@ -16,6 +17,7 @@ typedef enum {
     WARN = 3,
     ERROR = 4,
     FATAL = 5,
+    ALERT = 6,
 } LogLevel;
 
 typedef enum {
@@ -41,35 +43,65 @@ class ZenohInfo : public Msg {
 
     std::optional<std::string> zid;
     std::optional<std::string> what_am_i;
-    std::optional<std::string> peers;
+    std::vector<std::string> peers;
     std::optional<std::string> prefix;
-    std::optional<std::string> routers;
+    std::vector<std::string> routers;
     std::optional<std::string> connect;
     std::optional<std::string> listen;
     
 
-    JsonDocument serialize() const {
+    Bytes serialize() const {
         JsonDocument doc;
         JsonObject obj = doc["ZenohInfo"].to<JsonObject>();
         if (zid)  obj["zid"] = *zid;
         if (what_am_i)  obj["what_am_i"] = *what_am_i;
-        if (peers)  obj["peers"] = *peers;
+        {
+            JsonArray arr = obj["peers"].to<JsonArray>();
+        for (const auto& item : peers) {
+            arr.add(item);
+        }
+        }
         if (prefix)  obj["prefix"] = *prefix;
-        if (routers)  obj["routers"] = *routers;
+        {
+            JsonArray arr = obj["routers"].to<JsonArray>();
+        for (const auto& item : routers) {
+            arr.add(item);
+        }
+        }
         if (connect)  obj["connect"] = *connect;
         if (listen)  obj["listen"] = *listen;
-        return doc;
+        std::string str;
+        serializeJson(doc,str);
+        return Bytes(str.begin(),str.end());
     }
 
-    void deserialize(const JsonObject& obj) {
+    bool deserialize(const Bytes& bytes) {
+        JsonDocument obj;
+        auto err = deserializeJson(obj,bytes);
+        if ( err != DeserializationError::Ok || obj.is<JsonObject>() == false ) {
+            return false;
+        };        
         if (obj["zid"].is<std::string>() )  zid = obj["zid"].as<std::string>();
         if (obj["what_am_i"].is<std::string>() )  what_am_i = obj["what_am_i"].as<std::string>();
-        if (obj["peers"].is<std::string>() )  peers = obj["peers"].as<std::string>();
+        if (obj["peers"].is<JsonArray>()) {
+            JsonArray arr = obj["peers"].as<JsonArray>();
+            peers.clear();
+            for (JsonVariant v : arr) {
+                peers.push_back(v.as<std::string>());
+            }
+        }
         if (obj["prefix"].is<std::string>() )  prefix = obj["prefix"].as<std::string>();
-        if (obj["routers"].is<std::string>() )  routers = obj["routers"].as<std::string>();
+        if (obj["routers"].is<JsonArray>()) {
+            JsonArray arr = obj["routers"].as<JsonArray>();
+            routers.clear();
+            for (JsonVariant v : arr) {
+                routers.push_back(v.as<std::string>());
+            }
+        }
         if (obj["connect"].is<std::string>() )  connect = obj["connect"].as<std::string>();
         if (obj["listen"].is<std::string>() )  listen = obj["listen"].as<std::string>();
-        }
+        return true;
+    }
 
 };
 
@@ -79,59 +111,39 @@ class LogInfo : public Msg {
     inline const char *type_id() const override { return id; }; 
     static const uint32_t ID = 34678;
 
-    std::string src;
     std::optional<LogLevel> level;
     std::optional<std::string> message;
     std::optional<int32_t> error_code;
+    std::optional<std::string> file;
+    std::optional<int32_t> line;
     
 
-    JsonDocument serialize() const {
+    Bytes serialize() const {
         JsonDocument doc;
         JsonObject obj = doc["LogInfo"].to<JsonObject>();
-        obj["src"] = src;
         if (level)  obj["level"] = *level;
         if (message)  obj["message"] = *message;
         if (error_code)  obj["error_code"] = *error_code;
-        return doc;
+        if (file)  obj["file"] = *file;
+        if (line)  obj["line"] = *line;
+        std::string str;
+        serializeJson(doc,str);
+        return Bytes(str.begin(),str.end());
     }
 
-    void deserialize(const JsonObject& obj) {
-        
+    bool deserialize(const Bytes& bytes) {
+        JsonDocument obj;
+        auto err = deserializeJson(obj,bytes);
+        if ( err != DeserializationError::Ok || obj.is<JsonObject>() == false ) {
+            return false;
+        };        
         if (obj["level"].is<LogLevel>() )  level = obj["level"].as<LogLevel>();
         if (obj["message"].is<std::string>() )  message = obj["message"].as<std::string>();
         if (obj["error_code"].is<int32_t>() )  error_code = obj["error_code"].as<int32_t>();
-        }
-
-};
-
-class AlertInfo : public Msg {
-    public:
-    static constexpr const char *id = "AlertInfo";     
-    inline const char *type_id() const override { return id; }; 
-    static const uint32_t ID = 64124;
-
-    std::string src;
-    std::optional<LogLevel> level;
-    std::optional<std::string> message;
-    std::optional<int32_t> error_code;
-    
-
-    JsonDocument serialize() const {
-        JsonDocument doc;
-        JsonObject obj = doc["AlertInfo"].to<JsonObject>();
-        obj["src"] = src;
-        if (level)  obj["level"] = *level;
-        if (message)  obj["message"] = *message;
-        if (error_code)  obj["error_code"] = *error_code;
-        return doc;
+        if (obj["file"].is<std::string>() )  file = obj["file"].as<std::string>();
+        if (obj["line"].is<int32_t>() )  line = obj["line"].as<int32_t>();
+        return true;
     }
-
-    void deserialize(const JsonObject& obj) {
-        
-        if (obj["level"].is<LogLevel>() )  level = obj["level"].as<LogLevel>();
-        if (obj["message"].is<std::string>() )  message = obj["message"].as<std::string>();
-        if (obj["error_code"].is<int32_t>() )  error_code = obj["error_code"].as<int32_t>();
-        }
 
 };
 
@@ -141,28 +153,33 @@ class SysCmd : public Msg {
     inline const char *type_id() const override { return id; }; 
     static const uint32_t ID = 51983;
 
-    std::string dst;
     std::string src;
     std::optional<uint64_t> set_time;
     std::optional<bool> reboot;
     
 
-    JsonDocument serialize() const {
+    Bytes serialize() const {
         JsonDocument doc;
         JsonObject obj = doc["SysCmd"].to<JsonObject>();
-        obj["dst"] = dst;
         obj["src"] = src;
         if (set_time)  obj["set_time"] = *set_time;
         if (reboot)  obj["reboot"] = *reboot;
-        return doc;
+        std::string str;
+        serializeJson(doc,str);
+        return Bytes(str.begin(),str.end());
     }
 
-    void deserialize(const JsonObject& obj) {
-        
+    bool deserialize(const Bytes& bytes) {
+        JsonDocument obj;
+        auto err = deserializeJson(obj,bytes);
+        if ( err != DeserializationError::Ok || obj.is<JsonObject>() == false ) {
+            return false;
+        };        
         
         if (obj["set_time"].is<uint64_t>() )  set_time = obj["set_time"].as<uint64_t>();
         if (obj["reboot"].is<bool>() )  reboot = obj["reboot"].as<bool>();
-        }
+        return true;
+    }
 
 };
 
@@ -172,31 +189,39 @@ class SysInfo : public Msg {
     inline const char *type_id() const override { return id; }; 
     static const uint32_t ID = 10347;
 
-    std::string src;
+    std::optional<uint64_t> utc;
     std::optional<uint64_t> uptime;
     std::optional<uint64_t> free_heap;
     std::optional<uint64_t> flash;
     std::optional<std::string> cpu_board;
     
 
-    JsonDocument serialize() const {
+    Bytes serialize() const {
         JsonDocument doc;
         JsonObject obj = doc["SysInfo"].to<JsonObject>();
-        obj["src"] = src;
+        if (utc)  obj["utc"] = *utc;
         if (uptime)  obj["uptime"] = *uptime;
         if (free_heap)  obj["free_heap"] = *free_heap;
         if (flash)  obj["flash"] = *flash;
         if (cpu_board)  obj["cpu_board"] = *cpu_board;
-        return doc;
+        std::string str;
+        serializeJson(doc,str);
+        return Bytes(str.begin(),str.end());
     }
 
-    void deserialize(const JsonObject& obj) {
-        
+    bool deserialize(const Bytes& bytes) {
+        JsonDocument obj;
+        auto err = deserializeJson(obj,bytes);
+        if ( err != DeserializationError::Ok || obj.is<JsonObject>() == false ) {
+            return false;
+        };        
+        if (obj["utc"].is<uint64_t>() )  utc = obj["utc"].as<uint64_t>();
         if (obj["uptime"].is<uint64_t>() )  uptime = obj["uptime"].as<uint64_t>();
         if (obj["free_heap"].is<uint64_t>() )  free_heap = obj["free_heap"].as<uint64_t>();
         if (obj["flash"].is<uint64_t>() )  flash = obj["flash"].as<uint64_t>();
         if (obj["cpu_board"].is<std::string>() )  cpu_board = obj["cpu_board"].as<std::string>();
-        }
+        return true;
+    }
 
 };
 
@@ -206,7 +231,6 @@ class WifiInfo : public Msg {
     inline const char *type_id() const override { return id; }; 
     static const uint32_t ID = 15363;
 
-    std::string src;
     std::optional<std::string> ssid;
     std::optional<std::string> bssid;
     std::optional<int32_t> rssi;
@@ -217,10 +241,9 @@ class WifiInfo : public Msg {
     std::optional<std::string> netmask;
     
 
-    JsonDocument serialize() const {
+    Bytes serialize() const {
         JsonDocument doc;
         JsonObject obj = doc["WifiInfo"].to<JsonObject>();
-        obj["src"] = src;
         if (ssid)  obj["ssid"] = *ssid;
         if (bssid)  obj["bssid"] = *bssid;
         if (rssi)  obj["rssi"] = *rssi;
@@ -229,11 +252,17 @@ class WifiInfo : public Msg {
         if (channel)  obj["channel"] = *channel;
         if (gateway)  obj["gateway"] = *gateway;
         if (netmask)  obj["netmask"] = *netmask;
-        return doc;
+        std::string str;
+        serializeJson(doc,str);
+        return Bytes(str.begin(),str.end());
     }
 
-    void deserialize(const JsonObject& obj) {
-        
+    bool deserialize(const Bytes& bytes) {
+        JsonDocument obj;
+        auto err = deserializeJson(obj,bytes);
+        if ( err != DeserializationError::Ok || obj.is<JsonObject>() == false ) {
+            return false;
+        };        
         if (obj["ssid"].is<std::string>() )  ssid = obj["ssid"].as<std::string>();
         if (obj["bssid"].is<std::string>() )  bssid = obj["bssid"].as<std::string>();
         if (obj["rssi"].is<int32_t>() )  rssi = obj["rssi"].as<int32_t>();
@@ -242,7 +271,8 @@ class WifiInfo : public Msg {
         if (obj["channel"].is<int32_t>() )  channel = obj["channel"].as<int32_t>();
         if (obj["gateway"].is<std::string>() )  gateway = obj["gateway"].as<std::string>();
         if (obj["netmask"].is<std::string>() )  netmask = obj["netmask"].as<std::string>();
-        }
+        return true;
+    }
 
 };
 
@@ -252,28 +282,33 @@ class MulticastInfo : public Msg {
     inline const char *type_id() const override { return id; }; 
     static const uint32_t ID = 61310;
 
-    std::string src;
     std::optional<std::string> group;
     std::optional<int32_t> port;
     std::optional<uint32_t> mtu;
     
 
-    JsonDocument serialize() const {
+    Bytes serialize() const {
         JsonDocument doc;
         JsonObject obj = doc["MulticastInfo"].to<JsonObject>();
-        obj["src"] = src;
         if (group)  obj["group"] = *group;
         if (port)  obj["port"] = *port;
         if (mtu)  obj["mtu"] = *mtu;
-        return doc;
+        std::string str;
+        serializeJson(doc,str);
+        return Bytes(str.begin(),str.end());
     }
 
-    void deserialize(const JsonObject& obj) {
-        
+    bool deserialize(const Bytes& bytes) {
+        JsonDocument obj;
+        auto err = deserializeJson(obj,bytes);
+        if ( err != DeserializationError::Ok || obj.is<JsonObject>() == false ) {
+            return false;
+        };        
         if (obj["group"].is<std::string>() )  group = obj["group"].as<std::string>();
         if (obj["port"].is<int32_t>() )  port = obj["port"].as<int32_t>();
         if (obj["mtu"].is<uint32_t>() )  mtu = obj["mtu"].as<uint32_t>();
-        }
+        return true;
+    }
 
 };
 
@@ -283,28 +318,33 @@ class HoverboardInfo : public Msg {
     inline const char *type_id() const override { return id; }; 
     static const uint32_t ID = 59150;
 
-    std::string src;
     std::optional<int32_t> speed;
     std::optional<int32_t> direction;
     std::optional<int32_t> currentA;
     
 
-    JsonDocument serialize() const {
+    Bytes serialize() const {
         JsonDocument doc;
         JsonObject obj = doc["HoverboardInfo"].to<JsonObject>();
-        obj["src"] = src;
         if (speed)  obj["speed"] = *speed;
         if (direction)  obj["direction"] = *direction;
         if (currentA)  obj["currentA"] = *currentA;
-        return doc;
+        std::string str;
+        serializeJson(doc,str);
+        return Bytes(str.begin(),str.end());
     }
 
-    void deserialize(const JsonObject& obj) {
-        
+    bool deserialize(const Bytes& bytes) {
+        JsonDocument obj;
+        auto err = deserializeJson(obj,bytes);
+        if ( err != DeserializationError::Ok || obj.is<JsonObject>() == false ) {
+            return false;
+        };        
         if (obj["speed"].is<int32_t>() )  speed = obj["speed"].as<int32_t>();
         if (obj["direction"].is<int32_t>() )  direction = obj["direction"].as<int32_t>();
         if (obj["currentA"].is<int32_t>() )  currentA = obj["currentA"].as<int32_t>();
-        }
+        return true;
+    }
 
 };
 
@@ -314,28 +354,33 @@ class HoverboardCmd : public Msg {
     inline const char *type_id() const override { return id; }; 
     static const uint32_t ID = 58218;
 
-    std::string dst;
     std::optional<std::string> src;
     std::optional<int32_t> speed;
     std::optional<int32_t> direction;
     
 
-    JsonDocument serialize() const {
+    Bytes serialize() const {
         JsonDocument doc;
         JsonObject obj = doc["HoverboardCmd"].to<JsonObject>();
-        obj["dst"] = dst;
         if (src)  obj["src"] = *src;
         if (speed)  obj["speed"] = *speed;
         if (direction)  obj["direction"] = *direction;
-        return doc;
+        std::string str;
+        serializeJson(doc,str);
+        return Bytes(str.begin(),str.end());
     }
 
-    void deserialize(const JsonObject& obj) {
-        
+    bool deserialize(const Bytes& bytes) {
+        JsonDocument obj;
+        auto err = deserializeJson(obj,bytes);
+        if ( err != DeserializationError::Ok || obj.is<JsonObject>() == false ) {
+            return false;
+        };        
         if (obj["src"].is<std::string>() )  src = obj["src"].as<std::string>();
         if (obj["speed"].is<int32_t>() )  speed = obj["speed"].as<int32_t>();
         if (obj["direction"].is<int32_t>() )  direction = obj["direction"].as<int32_t>();
-        }
+        return true;
+    }
 
 };
 
@@ -349,18 +394,26 @@ class LpsInfo : public Msg {
     std::optional<std::string> msg;
     
 
-    JsonDocument serialize() const {
+    Bytes serialize() const {
         JsonDocument doc;
         JsonObject obj = doc["LpsInfo"].to<JsonObject>();
         if (direction)  obj["direction"] = *direction;
         if (msg)  obj["msg"] = *msg;
-        return doc;
+        std::string str;
+        serializeJson(doc,str);
+        return Bytes(str.begin(),str.end());
     }
 
-    void deserialize(const JsonObject& obj) {
+    bool deserialize(const Bytes& bytes) {
+        JsonDocument obj;
+        auto err = deserializeJson(obj,bytes);
+        if ( err != DeserializationError::Ok || obj.is<JsonObject>() == false ) {
+            return false;
+        };        
         if (obj["direction"].is<int32_t>() )  direction = obj["direction"].as<int32_t>();
         if (obj["msg"].is<std::string>() )  msg = obj["msg"].as<std::string>();
-        }
+        return true;
+    }
 
 };
